+++ package.js
@@ -24,6 +24,7 @@ Package.onTest( function( api ){
 
 function configure( api ){
     api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
+    api.use( 'accounts-base' );
     api.use( 'check' );
     api.use( 'ecmascript' );
     api.use( 'mongo' );